Add owner_id and sale_owner
